requirement.txt
absl-py==2.1.0
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
